[FIX] 카카오 로그인 수정

+++ app/src/main/java/kr/co/pureum/views/quest/QuestViewModel.kt
@@ -49,7 +49,7 @@ class QuestViewModel @Inject constructor(
         }
     }
 
-    fun getSentencesComplete(userId: Int) {
+    fun getSentencesComplete(userId: Long) {
         viewModelScope.launch {
             val res = repository.sentencesComplete(userId)
             //_todayKeywordLiveData.value = res.keyword.toString()
